backend/requirements.txt
django<3
psycopg2-binary==2.8.4
coverage==5.0.4
factory_boy==2.12.0
pillow==7.0.0
djangorestframework==3.11.0
markdown==3.2.1
django-rest-multiple-models==2.1.3
django-ipware==2.1.0
djangorestframework-simplejwt==4.6.0
djangorestframework-recaptcha==0.2.0
django-cors-headers==3.2.1
numpy==1.18.3
pandas==1.0.3
matplotlib==3.2.1
seaborn==0.10.1
requests==2.23.0
geopandas==0.7.0
django-rest-swagger
django-grappelli==2.14.2
sqlparse==0.4.1
moz-sql-parser==4.5.20343
pycrypto==2.6.1
cryptography==3.3.1
